requirements.txt
asgiref==3.8.1
attrs==23.2.0
Django==5.0.6
django-debug-toolbar==4.4.2
django-filter==24.2
djangorestframework==3.15.1
drf-nested-routers==0.94.1
drf-spectacular==0.27.2
Faker==25.5.0
flake8==7.0.0
inflection==0.5.1
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
mccabe==0.7.0
mysql-connector==2.2.9
mysql-connector-python==8.4.0
mysqlclient==2.2.4
pillow==10.3.0
pycodestyle==2.11.1
pyflakes==3.2.0
python-dateutil==2.9.0.post0
PyYAML==6.0.1
referencing==0.35.1
rpds-py==0.18.1
six==1.16.0
sqlparse==0.5.0
typing_extensions==4.12.1
uritemplate==4.1.1
